modify1

+++ functions_1.c
@@ -15,7 +15,7 @@ int print_unsigned(va_list types, char buffer[],
 	int flags, int width, int precision, int size)
 {
 	int i = BUFF_SIZE - 2;
-	unsigned long int num = va_arg(types, unsigned long int);
+	unsigned int num = va_arg(types, unsigned int);
 
 	num = convert_size_unsigned int (num, size);
